% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agromet_API.R
\name{get_data}
\alias{get_data}
\title{Retrieve data from the [AGROMET API](https://app.pameseb.be/fr/pages/api_call_test/).}
\usage{
get_data(user_token = Sys.getenv("AGROMET_API_V1_KEY"),
  table_name = "cleandata", sensors = "tsa", sid = "all",
  dfrom = Sys.Date() - 2, dto = Sys.Date() - 1, month_day = NULL,
  api_v = "v2", test.bool = FALSE)
}
\arguments{
\item{user_token}{A character specifying your user token. Default reads your .Renviron file}

\item{table_name}{A character specifying the database table you want to query.}

\item{sensors}{A character vector containing the names of the sensors you want to query}

\item{sid}{A character vector containing the ids of the stations you want to query}

\item{dfrom}{A character specifying the data from which you want to query data (YYYY-MM-DD)}

\item{dto}{A character specifying the data to which you want to query data (YYYY-MM-DD)}

\item{month_day}{A character specifying the month and day from which you want TMY data (MM-DD)}

\item{api_v}{A character specifying the version of the API you want to use ("v1" or "v2")}

\item{test.bool}{A boolean, TRUE if you want to query the test server}

\item{user_token}{A character specifying your user token}
}
\value{
A list containing the metadata in the first element and the queried data in the second element
}
\description{
Retrieve data from the [AGROMET API](https://app.pameseb.be/fr/pages/api_call_test/).
}
